refactor: tests + tooling — shared helpers, conftest fixtures (design 0000056 PR 3/5)
