+++ cafleet/pyproject.toml
@@ -31,14 +31,9 @@ include = [
 [dependency-groups]
 dev = [
     "pytest>=9.0.2",
-    "pytest-asyncio>=1.3.0",
     "ruff>=0.11.0",
     "ty>=0.0.26",
 ]
-
-[tool.pytest.ini_options]
-
-[tool.ty]
 
 [tool.ty.environment]
 python-version = "3.12"
