+++ cafleet/tests/test_alembic_0006_upgrade.py
@@ -1,27 +1,13 @@
 """Tests for Alembic migration ``0006_seed_administrator_agent`` (design 0000025 §C)."""
 
-import importlib.resources
 import json
 import uuid
-from datetime import UTC, datetime
 
 import pytest
 from alembic import command
-from alembic.config import Config
 from sqlalchemy import create_engine, text
 
-
-def _make_alembic_cfg(db_path) -> Config:
-    with importlib.resources.as_file(
-        importlib.resources.files("cafleet") / "alembic.ini"
-    ) as ini_path:
-        cfg = Config(str(ini_path))
-        cfg.set_main_option("sqlalchemy.url", f"sqlite:///{db_path}")
-        return cfg
-
-
-def _now_iso() -> str:
-    return datetime.now(UTC).isoformat()
+from tests._helpers import _make_alembic_cfg
 
 
 def _seed_session(
