+++ cafleet/tests/test_broker_member_activity.py
@@ -28,25 +28,14 @@ Design doc requirements exercised here:
 import uuid
 
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-import cafleet.db.engine  # noqa: F401 — registers PRAGMA listener globally
 from cafleet import broker
-from cafleet.db.models import Base
 from cafleet.tmux import DirectorContext
 
 
-@pytest.fixture
-def sync_sessionmaker():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    return sessionmaker(engine, expire_on_commit=False)
-
-
 @pytest.fixture(autouse=True)
-def _patch_broker(sync_sessionmaker, monkeypatch):
-    monkeypatch.setattr(broker, "get_sync_sessionmaker", lambda: sync_sessionmaker)
+def _autouse_broker(broker_session):
+    return broker_session
 
 
 def _bootstrap_session() -> tuple[str, str]:
