+++ cafleet/tests/test_broker_registry.py
@@ -5,56 +5,17 @@ import uuid
 
 import click
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-import cafleet.db.engine  # noqa: F401 — registers PRAGMA listener globally
 from cafleet import broker
 from cafleet.broker import ADMINISTRATOR_KIND
-from cafleet.db.models import Agent, Base
+from cafleet.db.models import Agent
 from cafleet.db.models import Session as SessionModel
-from cafleet.tmux import DirectorContext
-
-
-@pytest.fixture
-def sync_sessionmaker():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    return sessionmaker(engine, expire_on_commit=False)
-
-
-@pytest.fixture
-def _patch_broker(sync_sessionmaker, monkeypatch):
-    monkeypatch.setattr(broker, "get_sync_sessionmaker", lambda: sync_sessionmaker)
+from tests._broker_helpers import _create_session, _register_agent
 
 
 @pytest.fixture(autouse=True)
-def broker_session(sync_sessionmaker, _patch_broker):
-    return sync_sessionmaker
-
-
-def _create_session(label: str | None = None) -> dict:
-    return broker.create_session(
-        label=label,
-        director_context=DirectorContext(session="main", window_id="@3", pane_id="%0"),
-        coding_agent="claude",
-    )
-
-
-def _register_agent(
-    session_id: str,
-    name: str = "test-agent",
-    description: str = "A test agent",
-    skills: list[dict] | None = None,
-    placement: dict | None = None,
-) -> dict:
-    return broker.register_agent(
-        session_id=session_id,
-        name=name,
-        description=description,
-        skills=skills,
-        placement=placement,
-    )
+def _autouse_broker(broker_session):
+    return broker_session
 
 
 # --- create_session: broker.create_session(label=None) → dict with session_id, label, created_at ---
