+++ cafleet/tests/test_cli_member_send_input.py
@@ -8,52 +8,15 @@ from click.testing import CliRunner
 
 from cafleet import broker, tmux
 from cafleet.cli import cli
-
-DIRECTOR_ID = "11111111-1111-1111-1111-111111111111"
-MEMBER_ID = "22222222-2222-2222-2222-222222222222"
-OTHER_DIRECTOR_ID = "33333333-3333-3333-3333-333333333333"
-PANE_ID = "%7"
-MEMBER_NAME = "Claude-B"
-
-
-# Sentinel for ``_agent(placement=...)`` so callers can pass explicit None
-# (meaning "no placement row, exercise the missing-placement branch") without
-# being silently coerced back to a default valid placement.
-_UNSET: object = object()
-
-
-def _placement(
-    *,
-    director_agent_id: str = DIRECTOR_ID,
-    tmux_pane_id: str | None = PANE_ID,
-    coding_agent: str = "claude",
-) -> dict:
-    return {
-        "director_agent_id": director_agent_id,
-        "tmux_session": "main",
-        "tmux_window_id": "@3",
-        "tmux_pane_id": tmux_pane_id,
-        "coding_agent": coding_agent,
-        "created_at": "2026-04-16T08:00:00+00:00",
-    }
-
-
-def _agent(
-    *,
-    agent_id: str = MEMBER_ID,
-    name: str = MEMBER_NAME,
-    placement: dict | None | object = _UNSET,
-) -> dict:
-    resolved_placement = _placement() if placement is _UNSET else placement
-    return {
-        "agent_id": agent_id,
-        "name": name,
-        "description": "Test member",
-        "status": "active",
-        "registered_at": "2026-04-16T08:00:00+00:00",
-        "kind": "user",
-        "placement": resolved_placement,
-    }
+from tests._member_cli_helpers import (
+    DIRECTOR_ID,
+    MEMBER_ID,
+    MEMBER_NAME,
+    OTHER_DIRECTOR_ID,
+    PANE_ID,
+    _agent,
+    _placement,
+)
 
 
 @pytest.fixture
