+++ cafleet/tests/test_output_render_broadcast_summary.py
@@ -19,46 +19,14 @@ restoration is explicitly out of scope (the Director's clarification).
 """
 
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-import cafleet.db.engine  # noqa: F401 — registers PRAGMA listener globally
 from cafleet import broker
-from cafleet.db.models import Base
-from cafleet.tmux import DirectorContext
-
-
-@pytest.fixture
-def sync_sessionmaker():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    return sessionmaker(engine, expire_on_commit=False)
-
-
-@pytest.fixture
-def _patch_broker(sync_sessionmaker, monkeypatch):
-    monkeypatch.setattr(broker, "get_sync_sessionmaker", lambda: sync_sessionmaker)
+from tests._broker_helpers import _create_session, _register_agent
 
 
 @pytest.fixture(autouse=True)
-def broker_session(sync_sessionmaker, _patch_broker):
-    return sync_sessionmaker
-
-
-def _create_session() -> dict:
-    return broker.create_session(
-        label=None,
-        director_context=DirectorContext(session="main", window_id="@3", pane_id="%0"),
-        coding_agent="claude",
-    )
-
-
-def _register_agent(session_id: str, name: str) -> dict:
-    return broker.register_agent(
-        session_id=session_id,
-        name=name,
-        description=f"{name} description",
-    )
+def _autouse_broker(broker_session):
+    return broker_session
 
 
 # ---------------------------------------------------------------------------
