+++ cafleet/tests/test_session_bootstrap.py
@@ -5,16 +5,12 @@ from unittest.mock import Mock
 
 import click
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-import cafleet.db.engine  # noqa: F401 — registers PRAGMA listener globally
 from cafleet import broker
 from cafleet.broker import _is_administrator
 from cafleet.db.models import (
     Agent,
     AgentPlacement,
-    Base,
     Task,
 )
 from cafleet.db.models import (
@@ -23,21 +19,9 @@ from cafleet.db.models import (
 from cafleet.tmux import DirectorContext
 
 
-@pytest.fixture
-def sync_sessionmaker():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    return sessionmaker(engine, expire_on_commit=False)
-
-
-@pytest.fixture
-def _patch_broker(sync_sessionmaker, monkeypatch):
-    monkeypatch.setattr(broker, "get_sync_sessionmaker", lambda: sync_sessionmaker)
-
-
 @pytest.fixture(autouse=True)
-def broker_session(sync_sessionmaker, _patch_broker):
-    return sync_sessionmaker
+def _autouse_broker(broker_session):
+    return broker_session
 
 
 @pytest.fixture
