+++ cafleet/tests/test_webui_api_format.py
@@ -8,14 +8,16 @@ field mapping rather than parameterising on an accessor.
 """
 
 import pytest
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-import cafleet.db.engine  # noqa: F401 — registers PRAGMA listener globally
 from cafleet import broker, webui_api
-from cafleet.db.models import Base
 from cafleet.tmux import DirectorContext
 from cafleet.webui_api import _format_messages
+
+
+@pytest.fixture(autouse=True)
+def _autouse_broker(broker_session):
+    return broker_session
+
 
 _EXPECTED_KEYS = {
     "task_id",
@@ -30,23 +32,6 @@ _EXPECTED_KEYS = {
     "origin_task_id",
     "body",
 }
-
-
-@pytest.fixture
-def sync_sessionmaker():
-    engine = create_engine("sqlite:///:memory:")
-    Base.metadata.create_all(engine)
-    return sessionmaker(engine, expire_on_commit=False)
-
-
-@pytest.fixture
-def _patch_broker(sync_sessionmaker, monkeypatch):
-    monkeypatch.setattr(broker, "get_sync_sessionmaker", lambda: sync_sessionmaker)
-
-
-@pytest.fixture(autouse=True)
-def broker_session(sync_sessionmaker, _patch_broker):
-    return sync_sessionmaker
 
 
 def _create_session() -> dict:
